restructure chat components and remove unused files for improved organization
